:whale: 开源协议为 GPL2.0 [#99] 增加 msgId 特性，默认只在 request/response 模式下生效 [#102] 业务框架 BarSkeleton 类增加动态属性，方便扩展. [#104] 新增实验性特性-脉冲通讯方式 [#103] 业务框架新增 Runner 机制，增强扩展性、规范性 功能增强 功能增强. [#111] 新增文档解析、文档生成的控制选项 BrokerClient 添加 AwareInject

+++ net-bolt/bolt-external/src/main/java/com/iohao/game/bolt/broker/client/external/session/UserSessions.java
@@ -228,7 +228,7 @@ public class UserSessions {
      *
      * @return 当前在线人数
      */
-    public long countOnline() {
+    public int countOnline() {
         return this.channelGroup.size();
     }
